^ Refactor Oauth client
