--- app/Oauth/Providers/FlickrServiceProvider.php
@@ -1,39 +1,0 @@
-<?php
-/**
- * Copyright (c) 2020 JOOservices Ltd
- * @author Viet Vu <jooservices@gmail.com>
- * @package XGallery
- * @license GPL
- * @license http://opensource.org/licenses/gpl-license.php GNU Public License
- */
-
-namespace App\Oauth\Providers;
-
-use Illuminate\Support\ServiceProvider;
-
-/**
- * Class FlickrServiceProvider
- * @package App\OAuth\Providers
- */
-class FlickrServiceProvider extends ServiceProvider
-{
-    /**
-     * Register any application services.
-     *
-     * @return void
-     */
-    public function register()
-    {
-        //
-    }
-
-    /**
-     * Bootstrap any application services.
-     *
-     * @return void
-     */
-    public function boot()
-    {
-        //
-    }
-}
